Commit de Ronan Le Boulch : - Correction d'une erreur de double pour un sqrt
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mptk/trunk@1737 0f77665a-23fb-0310-bc23-e91fd8efc3e1

+++ src/plugin/contrib/lam/mdct_block_plugin.cpp
@@ -219,7 +219,7 @@ MP_Mdct_Block_Plugin_c::init_parameters (const unsigned long int setFilterLen,
 
 	/* initialize lapSize */
 	lapSize = setFilterLen/4;
-	scalingFactor = 1/sqrt(lapSize);
+	scalingFactor = 1/sqrt((double)lapSize);
 
 	/* Create the DCT object */
 	dct = (MP_DCT_Interface_c*)MP_DCT_Interface_c::init(numFilters);
